\name{coef.ICBayes}
\alias{coef.ICBayes}

\title{coef method for class ICBayes}
\description{extracts estimated regression coefficient(s)}
\usage{\method{coef}{ICBayes}(object, ...)}

\arguments{
  \item{object}{class ICBayes object}
  \item{\dots}{other arguments if any}
}
\value{a scalar or vector}