\name{ICBayes-package}
\alias{ICBayes-package}
\docType{package}
\title{Bayesian Semiparametric Models for Interval-Censored Data}
\description{
This package contains functions to fit several survival regression models 
(including the proportional hazard model, the proportional odds model, and the probit model) for interval-censored data under Bayesian framework. 
Estimations are available for both regression coefficients and survival functions.The Bayesian model selection criterion log pseudo marginal 
likelihood (LPML) is computed. 
}
\details{
\tabular{ll}{
Package: \tab ICBayes\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-12-24\cr
License: \tab GPL>=2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Chun Pan, Bo Cai, Lianming Wang, and Xiaoyan Lin

Maintainer: Chun Pan \email{chunpan2003@hotmail.com}
}
