\name{logLik.ICBayes}
\alias{logLik.ICBayes}
\title{logLik method for class ICBayes}
\description{log-likelihood of the observed interval-censored data estimated by log pseudo marginal likelihood}
\usage{\method{logLik}{ICBayes}(object, ...)}

\arguments{
  \item{object}{class ICBayes object}
  \item{\dots}{other arguments if any}
}

\value{a scalar}
