\name{summary.ICBayes}
\alias{summary.ICBayes}
\title{summary method for class ICBayes}
\description{present output from function \code{\link{ICBayes}}}
\usage{\method{summary}{ICBayes}(object, ...)}

\arguments{
  \item{object}{class ICBayes object}
  \item{\dots}{other arguments if any}
}

\value{
an object of class \code{ICBayes}.
}
